19551: Fixes an issue where tracefile label names are in quotes (#92)
Running tracefiles with label names surrounded in quotes was not
executing the labels correctly.

My change here moves and renames `Platform_SplitArgString` to
`StringManipulation::SplitArgString` (and moves it from
PlatformSpecific.h/cpp to StringManipulation.h/cpp)

Additionally, a boolean flag is added to this method named `greedy`. As
the name suggests, if `greedy` is True, then the method returns the full
vector of args. If it is False, then a vector with only the first arg is
returned and the argument string is modified to have that first word cut
out.

Then, `RemoveFirstWord` is renamed to `RemoveFirstToken` and modified to
use this new modified version of `SplitArgString` with the `greedy` flag
set to False to remove the first word, then returns an empty string if
no string is returned.

---------

Co-authored-by: J. Caleb Wherry <337871+calebwherry@users.noreply.github.com>
Co-authored-by: howsoRes <144272317+howsoRes@users.noreply.github.com>