24630: load/store to/from memory, entities, MAJOR
Add new LoadEntityFromMemory()/StoreEntityToMemory() external API calls
that take the entity contents from a pair of void *, size_t rather than
a file.

For all of the Load/StoreEntity APIs, add a pair of char **, size_t
parameters that identify a path to a non-root entity.  This allows
loading or storing something nested potentially deep within the
entity tree.

Existing calls to LoadEntity() and StoreEntity() need to add the
additional entity-path parameters, which will usually be nullptr, 0.
If LoadEntity is passed a non-null entity path, then its return value
will include an entity path as well, which needs to be separately
freed.